+++ crates/reporter/Cargo.toml
@@ -11,8 +11,9 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
-serde      = { workspace = true }
-serde_json = { workspace = true }
+gethostname = { workspace = true }
+serde       = { workspace = true }
+serde_json  = { workspace = true }
 
 [dev-dependencies]
 insta             = { workspace = true }
